requirements.txt
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
dj-database-url==0.5.0
Django==2.1.7
django-appconf==1.0.3
django-axes==5.2.2
django-cors-headers==3.2.1
django-countries==5.5
django-crispy-forms==1.8.1
django-email==0.1.10
django-environ==0.4.5
django-ipware==2.1.0
django-redis==4.11.0
djangorestframework==3.11.0
environ==1.0
httpie==2.0.0
idna==2.9
loadtest==0.0.2
pep8==1.7.1
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.2
python-memcached==1.59
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
slugify==0.0.1
unicode==2.7
unicodedata2==12.1.0
Unidecode==1.1.1
urllib3==1.25.8
